Updated Red Coins in Coin Blocks
This PR adjusts the conditions of the Red Coin when they're contained in a Coin Brick Block.
NOTE: In SMB Deluxe, the Red Coin appears exactly on the 10th hit of a Coin Brick Block, but I decided to have it appear after you hit the block 10 times or over to be looser and to rack up more score.
- When hitting a Coin Brick Block after some time, it produces 1 coin after hitting again instead of 2
- The Red Coin will only pop out of Coin Brick Blocks after being hit at least 9 times prior to the timeout. On the 10th hit or greater, the coin will pop out.